added a return value to indicate success of
2002-05-15  Sven Neumann  <sven@gimp.org>

	* app/config/gimpconfig.[ch]: added a return value to indicate
	success of GimpConfigInterface::serialize. Added a GimpConfigError
	value for write errors and check the return value of all write()
	calls. Allow to specify file header and footer when calling
	gimp_config_serialize().

	* app/config/gimpconfig-serialize.[ch]
	* app/config/gimprc.c: changed accordingly.

	* app/core/gimp-documents.[ch]: new files with functions to handle
	gimp->documents.

	* app/core/gimpdocuments.[ch]: removed ...

	* app/core/gimpdocumentlist.[ch]: ... and added under this name.
	Implements a generic document list.

	* app/core/gimp-parasites.[ch]: new files with functions to handle
	gimp->parasites.

	* app/core/gimpparasite.[ch]: removed.

	* app/core/gimpparasitelist.[ch]: nothing more than a generic
	parasite list.

	* app/undo.c
	* app/core/Makefile.am
	* app/core/core-types.h
	* app/core/gimp.c
	* app/core/gimpdrawable.c
	* app/core/gimpimage-merge.c
	* app/core/gimpimage.c
	* app/core/gimpitem.c
	* app/file/file-open.c
	* app/file/file-save.c
	* app/gui/file-open-dialog.c
	* app/pdb/parasite_cmds.c
	* tools/pdbgen/pdb/parasite.pdb: changed accordingly.

+++ libgimpconfig/gimpconfig-iface.c
@@ -47,7 +47,7 @@
 
 static void  gimp_config_iface_init (GimpConfigInterface  *gimp_config_iface);
 
-static void      gimp_config_iface_serialize    (GObject  *object,
+static gboolean  gimp_config_iface_serialize    (GObject  *object,
                                                  gint      fd);
 static gboolean  gimp_config_iface_deserialize  (GObject  *object,
                                                  GScanner *scanner);
@@ -94,11 +94,11 @@ gimp_config_iface_init (GimpConfigInterface *gimp_config_iface)
   gimp_config_iface->equal       = gimp_config_iface_equal;
 }
 
-static void
+static gboolean
 gimp_config_iface_serialize (GObject *object,
                              gint     fd)
 {
-  gimp_config_serialize_properties (object, fd);
+  return gimp_config_serialize_properties (object, fd);
 }
 
 static gboolean
@@ -163,6 +163,8 @@ gimp_config_iface_equal (GObject *a,
  * gimp_config_serialize:
  * @object: a #GObject that implements the #GimpConfigInterface.
  * @filename: the name of the file to write the configuration to.
+ * @header: optional file header (should be a comment)
+ * @footer: optional file footer (should be a comment)
  * @error:
  * 
  * Serializes the object properties of @object to the file specified
@@ -175,9 +177,12 @@ gimp_config_iface_equal (GObject *a,
 gboolean
 gimp_config_serialize (GObject      *object,
                        const gchar  *filename,
+                       const gchar  *header,
+                       const gchar  *footer,
                        GError      **error)
 {
   GimpConfigInterface *gimp_config_iface;
+  gboolean             success = TRUE;
   gchar               *tmpname;
   gint                 fd;
 
@@ -196,32 +201,48 @@ gimp_config_serialize (GObject      *object,
   if (fd == -1)
     {
       g_set_error (error, 
-                   GIMP_CONFIG_ERROR, GIMP_CONFIG_ERROR_FILE, 
-                   _("Failed to generate temporary file for '%s': %s"),
+                   GIMP_CONFIG_ERROR, GIMP_CONFIG_ERROR_OPEN, 
+                   _("Failed to create temporary file for '%s': %s"),
                    filename, g_strerror (errno));
       g_free (tmpname);
       return FALSE;
     }
 
-  gimp_config_iface->serialize (object, fd);
+  if (header)
+    success = (write (fd, header, strlen (header)) != -1);
+
+  if (success)
+    success = gimp_config_iface->serialize (object, fd);
+
+  if (success && footer)
+    success = (write (fd, footer, strlen (footer)) != -1);
+
+  if (! success)
+    {
+      g_set_error (error,
+                   GIMP_CONFIG_ERROR, GIMP_CONFIG_ERROR_WRITE,
+                   _("Error when writing to file '%s': %s"),
+                   tmpname, g_strerror (errno));
+
+      unlink (tmpname);
+    }
 
   close (fd);
 
-  if (rename (tmpname, filename) == -1)
+  if (success && rename (tmpname, filename) == -1)
     {
       g_set_error (error, 
-                   GIMP_CONFIG_ERROR, GIMP_CONFIG_ERROR_FILE, 
-                   _("Failed to rename temporary file '%s' to '%s': %s"),
-                   tmpname, filename, g_strerror (errno));
-      
+                   GIMP_CONFIG_ERROR, GIMP_CONFIG_ERROR_OPEN, 
+                   _("Failed to create file '%s': %s"),
+                   filename, g_strerror (errno));
+
       unlink (tmpname);
-      g_free (tmpname);
-      return FALSE;
+      success = FALSE;
     }
 
   g_free (tmpname);
 
-  return TRUE;
+  return success;
 }
 
 /**
@@ -262,7 +283,7 @@ gimp_config_deserialize (GObject      *object,
       g_set_error (error,
                    GIMP_CONFIG_ERROR, 
                    (errno == ENOENT ?
-                    GIMP_CONFIG_ERROR_FILE_ENOENT : GIMP_CONFIG_ERROR_FILE),
+                    GIMP_CONFIG_ERROR_ENOENT : GIMP_CONFIG_ERROR_OPEN),
                    _("Failed to open file: '%s': %s"),
                    filename, g_strerror (errno));
       return FALSE;
